+++ example/android/javalib/4-sum-lib-java/lib/src/main/java/com/sumlib/Sum.java
@@ -1,14 +1,12 @@
 package com.sumlib;
 
 public class Sum {
-    public static int apply(int[] numbers) {
+  public static int apply(int[] numbers) {
 
-        int sum = 0;
-        for (int i = 0; i < numbers.length; i++) {
-            sum += numbers[i];
-        }
-        return sum;
-
+    int sum = 0;
+    for (int i = 0; i < numbers.length; i++) {
+      sum += numbers[i];
     }
-
+    return sum;
+  }
 }
