+++ scalalib/src/mill/javalib/android/AndroidAppModule.scala
@@ -1,6 +1,5 @@
 package mill.javalib.android
 
-import coursier.Repository
 import mill._
 import mill.scalalib._
 import mill.api.{Logger, PathRef, internal}
@@ -52,86 +51,9 @@ object AndroidLintReportFormat extends Enumeration {
  * [[https://developer.android.com/studio Android Studio Documentation]]
  */
 @mill.api.experimental
-trait AndroidAppModule extends JavaModule {
+trait AndroidAppModule extends AndroidModule {
 
   private val parent: AndroidAppModule = this
-
-  // https://cs.android.com/android-studio/platform/tools/base/+/mirror-goog-studio-main:build-system/gradle-core/src/main/java/com/android/build/gradle/internal/tasks/D8BundleMainDexListTask.kt;l=210-223;drc=66ab6bccb85ce3ed7b371535929a69f494d807f0
-  private val mainDexPlatformRules = Seq(
-    "-keep public class * extends android.app.Instrumentation {\n" +
-      "  <init>(); \n" +
-      "  void onCreate(...);\n" +
-      "  android.app.Application newApplication(...);\n" +
-      "  void callApplicationOnCreate(android.app.Application);\n" +
-      "}",
-    "-keep public class * extends android.app.Application { " +
-      "  <init>();\n" +
-      "  void attachBaseContext(android.content.Context);\n" +
-      "}",
-    "-keep public class * extends android.app.backup.BackupAgent { <init>(); }",
-    "-keep public class * extends android.test.InstrumentationTestCase { <init>(); }"
-  )
-
-  private val compiledResourcesDirName = "compiled-resources"
-  private val rClassDirName = "RClass"
-
-  protected val debugKeyStorePass = "mill-android"
-  protected val debugKeyAlias = "mill-android"
-  protected val debugKeyPass = "mill-android"
-
-  /**
-   * Adds "aar" to the handled artifact types.
-   */
-  override def artifactTypes: T[Set[coursier.Type]] =
-    Task { super.artifactTypes() + coursier.Type("aar") }
-
-  override def repositoriesTask: Task[Seq[Repository]] = Task.Anon {
-    super.repositoriesTask() :+ AndroidSdkModule.mavenGoogle
-  }
-
-  override def sources: T[Seq[PathRef]] = Task.Sources(moduleDir / "src/main/java")
-
-  /**
-   * Provides access to the Android SDK configuration.
-   */
-  def androidSdkModule: ModuleRef[AndroidSdkModule]
-
-  /**
-   * Provides os.Path to an XML file containing configuration and metadata about your android application.
-   */
-  def androidManifest: Task[PathRef] = Task.Source("src/main/AndroidManifest.xml")
-
-  /**
-   * Name of the release keystore file. Default is not set.
-   */
-  def androidReleaseKeyName: T[Option[String]] = Task { None }
-
-  // TODO alias, keystore pass and pass below are sensitive credentials and shouldn't be leaked to disk/console.
-  // In the current state they are leaked, because Task dumps output to the json.
-  // Should be fixed ASAP.
-
-  /**
-   * Name of the key alias in the release keystore. Default is not set.
-   */
-  def androidReleaseKeyAlias: T[Option[String]] = Task { None }
-
-  /**
-   * Password for the release key. Default is not set.
-   */
-  def androidReleaseKeyPass: T[Option[String]] = Task { None }
-
-  /**
-   * Password for the release keystore. Default is not set.
-   */
-  def androidReleaseKeyStorePass: T[Option[String]] = Task { None }
-
-  /**
-   * Default os.Path to the keystore file, derived from `androidReleaseKeyName()`.
-   * Users can customize the keystore file name to change this path.
-   */
-  def androidReleaseKeyPath: T[Option[PathRef]] = Task {
-    androidReleaseKeyName().map(name => PathRef(moduleDir / name))
-  }
 
   /**
    * Specifies the file format(s) of the lint report. Available file formats are defined in AndroidLintReportFormat,
@@ -162,113 +84,9 @@ trait AndroidAppModule extends JavaModule {
    */
   def androidLintArgs: T[Seq[String]] = Task { Seq.empty[String] }
 
-  /**
-   * This setting defines which Android API level your project compiles against. This is a required property for
-   * Android builds.
-   *
-   * It determines what Android APIs are available during compilation - your code can only use APIs from this level
-   * or lower.
-   *
-   * See [[https://developer.android.com/guide/topics/manifest/uses-sdk-element.html#ApiLevels]] for more details.
-   */
-  def androidCompileSdk: T[Int]
-
-  /**
-   * The minimum SDK version to use. Default is 1.
-   *
-   * See [[https://developer.android.com/guide/topics/manifest/uses-sdk-element.html#min]] for more details.
-   */
-  def androidMinSdk: T[Int] = 1
-
-  /**
-   * The target SDK version to use. Default is equal to the [[androidCompileSdk]] value.
-   *
-   * See [[https://developer.android.com/guide/topics/manifest/uses-sdk-element.html#target]] for more details.
-   */
-  def androidTargetSdk: T[Int] = androidCompileSdk
-
-  /**
-   * The version name of the application. Default is "1.0".
-   *
-   * See [[https://developer.android.com/studio/publish/versioning]] for more details.
-   */
-  def androidVersionName: T[String] = "1.0"
-
-  /**
-   * Version code of the application. Default is 1.
-   *
-   * See [[https://developer.android.com/studio/publish/versioning]] for more details.
-   */
-  def androidVersionCode: T[Int] = 1
-
-  /**
-   * Controls debug vs release build type. Default is `true`, meaning debug build will be generated.
-   *
-   * This option will probably go away in the future once build variants are supported.
-   */
-  def androidIsDebug: T[Boolean] = true
-
-  def aarPath: T[PathRef] = Task{
+  def aarPath: T[PathRef] = Task {
     val aarPath = T.dest / "library.aar"
     PathRef(aarPath)
-  }
-
-
-  /**
-   * Extracts JAR files and resources from AAR dependencies.
-   */
-  def androidUnpackArchives: T[Seq[UnpackedDep]] = Task {
-    // The way Android is handling configurations for dependencies is a bit different from canonical Maven: it has
-    // `api` and `implementation` configurations. If dependency is `api` dependency, it will be exposed to consumers
-    // of the library, but if it is `implementation`, then it won't. The simplest analogy is api = compile,
-    // implementation = runtime, but both are actually used for compilation and packaging of the final DEX.
-    // More here https://docs.gradle.org/current/userguide/java_library_plugin.html#sec:java_library_separation.
-    //
-    // In Gradle terms using only `resolvedRunIvyDeps` won't be complete, because source modules can be also
-    // api/implementation, but Mill has no such configurations.
-    val aarFiles = (super.compileClasspath() ++ super.resolvedRunIvyDeps())
-      .map(_.path)
-      .filter(_.ext == "aar")
-      .distinct
-
-    // TODO do it in some shared location, otherwise each module is doing the same, having its own copy for nothing
-    extractAarFiles(aarFiles, Task.dest)
-  }
-
-  final def extractAarFiles(aarFiles: Seq[os.Path], taskDest: os.Path): Seq[UnpackedDep] = {
-    aarFiles.map(aarFile => {
-      val extractDir = taskDest / aarFile.baseName
-      os.unzip(aarFile, extractDir)
-      val name = aarFile.baseName
-
-      def pathOption(p: os.Path): Option[PathRef] = if (os.exists(p)) {
-        Some(PathRef(p))
-      } else None
-
-      val classesJar = pathOption(extractDir / "classes.jar")
-      val proguardRules = pathOption(extractDir / "proguard.txt")
-      val resources = pathOption(extractDir / "res")
-      val manifest = pathOption(extractDir / "AndroidManifest.xml")
-      val lintJar = pathOption(extractDir / "lint.jar")
-      val metaInf = pathOption(extractDir / "META-INF")
-      val nativeLibs = pathOption(extractDir / "jni")
-      val baselineProfile = pathOption(extractDir / "baseline-prof.txt")
-      val stableIdsRFile = pathOption(extractDir / "R.txt")
-      val publicResFile = pathOption(extractDir / "public.txt")
-      UnpackedDep(
-        name,
-        classesJar,
-        proguardRules,
-        resources,
-        manifest,
-        lintJar,
-        metaInf,
-        nativeLibs,
-        baselineProfile,
-        stableIdsRFile,
-        publicResFile
-      )
-    })
   }
 
   /**
@@ -308,7 +126,6 @@ trait AndroidAppModule extends JavaModule {
       .distinct
     transformedAarFilesToJar ++ jarFiles
   }
-
 
   /**
    * Replaces AAR files in classpath with their extracted JARs.
